add paper-scale experiment results (A100) and update experiment log
- Run all 7 experiments on Google Colab A100 (30 rounds, 10 clients,
  200 samples/client, seed 42): BatteryLoRA, HomLoRA r=8, HomLoRA r=32,
  HetLoRA, and 3 ablation policies (binary, random, continuous)
- Update EXPERIMENT_LOG.md with complete results tables, per-client
  rank adaptation detail, convergence analysis, and key findings
- Fix client.py for trl 0.12 compatibility (TrainingArguments, tokenizer
  param, batched formatting, bf16, remove_unused_columns)
- Fix model_utils.py: torch_dtype for transformers 4.48
- Fix data_utils.py: handle both batched and unbatched formatting
- Add run_all.py: experiment runner with per-experiment Drive backup,
  skip-if-done resumability, and result summary table
- Remove private research planning docs from tracking (gitignored)

+++ battery-lora/model_utils.py
@@ -106,7 +106,7 @@ def load_base_model(model_cfg: ModelConfig):
     model = AutoModelForCausalLM.from_pretrained(
         model_cfg.name,
         quantization_config=quant_config,
-        dtype=dtype,
+        torch_dtype=dtype,
         device_map=device_map,
         trust_remote_code=True,
     )
